feat: front/delete task - 컨펌 후 삭제 - 삭제 성공 후 새로고침

+++ front-end/src/components/taskForm.ts
@@ -1,6 +1,5 @@
 import { getTaskCount, createTask, getMonthlyTaskList, getDaliyTaskList } from "../services/taskService.js";
 import { getSelectedDate } from "./calenderForm.js";
-import { renderTasks } from "../utils/taskRenderUtils.js";
 
 export async function getMonthlyTaskProcess(page: number) {
     let seletedDate: number[] = getSelectedDate();
@@ -22,32 +21,44 @@ function formatDate(year: number, month: number, day: number): string {
     return date.toISOString().split("T")[0];
 }
 
-const createBtn = document.getElementById('create');
+const createBtn = document.getElementById('create') as HTMLButtonElement;
+
+const taskInput = document.getElementById('task-input') as HTMLInputElement;
+taskInput.addEventListener("input", function () {
+    createBtn.disabled = taskInput.value.trim() === "";
+});
+
 if (createBtn) {
     createBtn.addEventListener('click', async () => {
+        const content = document.getElementById('task-input') as HTMLInputElement;
+
         try {
-            const content = document.getElementById('task-input') as HTMLInputElement;
-            const date = document.getElementById('input-date') as HTMLElement;
-            let splitDate = date.innerText.split(". ");
+            const isConfirmed = confirm("등록하시겠습니까?");
+            if (isConfirmed) {
+                const date = document.getElementById('input-date') as HTMLElement;
+                let splitDate = date.innerText.split(". ");
 
-            const task = {
-                content: content.value,
-                date: formatDate(Number(splitDate[0]), Number(splitDate[1]), Number(splitDate[2]))
-            };
+                const task = {
+                    content: content.value,
+                    date: formatDate(Number(splitDate[0]), Number(splitDate[1]), Number(splitDate[2]))
+                };
 
-            await createTask(task)
-            .then((response: any) => {
-                if (response.status === 200) {
-                    alert("등록이 완료 되었습니다.");
-                    localStorage.setItem("year", splitDate[0]);
-                    localStorage.setItem("month", splitDate[1]);
-                    window.location.reload();
-                } else {
-                    alert(response.message);
-                }
-            })
-        } catch (error) {
-            console.log(error);
+                await createTask(task)
+                .then((response: any) => {
+                    if (response.status === 200) {
+                        alert("등록이 완료되었습니다.");
+                        localStorage.setItem("year", splitDate[0]);
+                        localStorage.setItem("month", splitDate[1]);
+                        window.location.reload();
+                    } else {
+                        alert(response.message);
+                    }
+                })
+            } else {
+                content.value = "";
+            }
+        } catch (error: any) {
+            console.log(error.message);
         }
     });
 }
